added classes necessary to fetch 5 day forecast

+++ app/src/main/java/soutvoid/com/DsrWeatherApp/domain/rain/Rain.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class Rain(
         @SerializedName("3h")
-        val rainVolume : Int
+        val rainVolume : Double
 )
