+++ app/src/main/java/soutvoid/com/DsrWeatherApp/domain/snow/Snow.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class Snow(
         @SerializedName("3h")
-        val snowVolume : Int
+        val snowVolume : Double
 )
